Move test from array_functions_test to functions_test.go
